requirements.txt
discord.py
flake8
pytest
pymongo
discord_components
logger
python-dotenv